+++ CHANGELOG.md
@@ -384,6 +384,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Version 1.17.0/0.38b0 (2023-03-22)
 
+- Fix opentelemetry-instrumentation-celery warnings when using kombu.Exchange and kombu.Queue
+  ([#1195](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/1196))
+
 ### Added
 
 - Add connection attributes to sqlalchemy connect span